Port (without fully testing) the inline shell from PengOS class to its own 'real' executable
